Atualização das funcionalidades da tela do projeto.

+++ backend/src/models/projects/relationship/projects-members/ProjectsMembersService.js
@@ -16,6 +16,7 @@ class ProjectMembersService extends Service {
 	/**
 	 * Retorna os membros de um projeto.
 	 *
+	 * @param {Object} pageOptions { PAGE: number, PAGE_SIZE: number }
 	 * @param {number} projectId
 	 * @returns {Page}
 	 */
@@ -42,9 +43,7 @@ class ProjectMembersService extends Service {
 	 * @returns {boolean}
 	 */
 	async addUserAsMemberOfProject(userId, projectId) {
-		return await this.#ProjectsMembersRepository.addUserAsMemberOfProject(
-			new ProjectsMembersDTO({ userId, projectId })
-		);
+		return await this.#ProjectsMembersRepository.addUserAsMemberOfProject(new ProjectsMembersDTO({ userId, projectId }));
 	}
 
 	/**
@@ -55,9 +54,7 @@ class ProjectMembersService extends Service {
 	 * @returns {boolean}
 	 */
 	async removeUserFromMembersOfProject(userId, projectId) {
-		return await this.#ProjectsMembersRepository.removeUserFromMembersOfProject(
-			new ProjectsMembersDTO({ userId, projectId })
-		);
+		return await this.#ProjectsMembersRepository.removeUserFromMembersOfProject(new ProjectsMembersDTO({ userId, projectId }));
 	}
 
 	/**
@@ -68,9 +65,7 @@ class ProjectMembersService extends Service {
 	 * @returns {boolean}
 	 */
 	async isUserMemberOfProject(userId, projectId) {
-		return await this.#ProjectsMembersRepository.isUserMemberOfProject(
-			new ProjectsMembersDTO({ userId, projectId })
-		);
+		return await this.#ProjectsMembersRepository.isUserMemberOfProject(new ProjectsMembersDTO({ userId, projectId }));
 	}
 }
